split target action detail
